+++ bundle/CHANGELOG.md
@@ -1,5 +1,21 @@
 ### Changelog
 
+##### 0.4.2
+- Remove `BepinProcess` to resolve [server side process name issue](https://github.com/kruftt/SmoothArmorScaling/issues/1).
+
+##### 0.4.1
+- Fix readme typo
+
+#### 0.4.0
+- Add `damageTaken` multiplier
+- Add `smoothArmorScaling` toggle
+- Adjust granularity of `armorEffectiveness` to 0.01
+- Adjust range of gear multiplier
+
+#### 0.3.0
+- Reinclude `armorEffectiveness` alongside armor settings.
+- Update documentation
+
 #### 0.2.0
 - Remove `armorEffectiveness`
 - Add player and gear armor configuration options